fix imports from package and a lot of linting
